Fix a crash when zooming local tiles (sulla) (#4424)
When zooming in local tiles we call `do_layout` to calculate the new
size of the view buffer (as well as other things). however, we need to
call `redraw_screen` to actually resize the view buffer before next
rendering or the game will crash. We were calling `mprf` before
`redraw_screen` which can lead to a force more which can sometimes
result in rendering.

This change maybe should be applied to earlier versions to, as in at
least version 0.32 zooming in map view mode causes force mores by
default.