Rework habitats
Store habitats as a bit mask of the different terrain categories
monsters of that habitat can survive on. This lets us easily check of
one habitat is more restrictive than another or if a monster can survive
on a category of terrain e.g. land, water, shallow water or deep water
etc.

[Committer's note: Fixed a minor bug with wizmode listing of monster
 habitats, removed unused random_monster_at_grid() function and simplified
 code supporting it.]