Reduce excessively high Necropolis chances, hopefully
The portal was missing a `chance_foo` tag as the other portals use, which
was resulting in much higher chances to see a Necropolis than CHANCE
should be providing. There's some worrying issues with objstat's feature
count reflecting very different odds from mapstat on whether or not adding
the tag does anything meaningful (possibly from not regenerating the .des
cache properly?), and whether or not the `chance_foo` tag properly works
to avoid rerolling into the same vault, but this does immediately help the
problem at hand by moving from 100% chance to place in D over 100
iterations to ~29% over 100 iterations.