Remove most autogenerate resistance fakemuts for forms
Now that the form properties screen lists all of these in a clear and
concise way, I don't think this is needed. And the current patchwork of
special cases (which existed before the new properties screen) claims a lot
of slightly-inaccurate things about different forms (like claiming a
werewolf's rC is from having an 'undead body') that are extremely not
future-proof for any form changes and additions.