Allow simulacra to swim
They're made of ice and ice floats!

(More pertinently, this was a mild issue with Sculpt Simulacrum, where the
blocks of ice it made *did* float, but the simulacra did not, sometimes
resulting in the blocks of ice hatching into nothing in confined spaces.)

For the most part, this should have limited effect on monster balance (the
notable simulacra used in Ice Caves generally could already cross water
due to their size or flight anyway).