Move sky beasts out of Y and into F
Nobody previously knew what a sky beast was, beyond "like an ice beast
that flies", itself a loose generic fantasy monster monster that was moved
somewhat questionably onto the 'Y' glyph alongside the other yaks and
elephants for being a simple non-carnivorous quadruped long ago. The sky
beast tile has moved from a vaguely-legged cloud (f20cd82), to a crackling
coatl (46161bd), to a very abstract cloudy ray (4a87d9e), to the current
feathered manta ray (f2ebe9a). The last of these as finally made it
somewhat coherent to parse also unfortunately clashes with its original
status long ago in bodyshape as derived undead tiles are assigned to it,
with it currently being shown as a winged quadruped if undead despite only
very briefly having legs compared to the last fifteen years.

With F mostly established as 'Frogs and 'F'ish (electric eels will be
addressed eventually....), I am taking the opportunity to both make its
glyph and its derived undead shape a fish, as manta rays are. This is
possibly pushing the 'F'ish part of the glyph a little (mostly for being
the only one of them not fast), but it hopefully makes substantially more
sense than leaving it in the same space as yaks and elephants.