Let solar ember charge actually decay
It was intended that the 'every second hit' property of solar flare
wouldn't maintain charge for long times (so there was no incentive to ever
'charge' it in advance), but apparently ENCH_SPELL_CHARGED did not decay
at all. Now it should properly only last a few turns (enough to reposition
mid-fight, but not to hold the charge for an extended time).