Tweak Shapeshifter starting kit (various)
Give them back their potion of lignification, which serves a different role
than flux baubles (since the latter isn't helpful on D:1).

Reduce the number of flux baubles from 4->3 (but increase the duration of
each by ~12%).

And finally, let them start with 2 Fighting skill instead of 1 (which is
pretty standard for backgrounds we expect to start the game by hitting
things.)