Better handle melded slot-granting items on floor transition (Zeldest)
The block of code in tags.cc that handles auto-unequipping items the game
determines should not be able to be equipped (primarily intended as save
compat, to deal with items or species slot restrictions being changed
between versions, but which happens to also get called between floors) was
overzealous in some situations.

For instance, if the player wears an unrand that grants additional item
slots, and transforms into a form which melds it (thus 'removing' those
slots), the items in those slots are simply melded until they untransform.
But upon loading the save (such as by changing floors), the game would not
see the slot-granting item, and thus think the player was wearing too many
items and remove some of them.

Now, when doing save cleanup, the game should count all slot-granting
unrands as if they were unmelded.