Refactor summon space checks, add a few more, fix oddities
This refactors the checks for 'enough room to summon a monster' added by
f2a3531c2594a0d with the intent of reducing boilerplate, by wrapping the
error message and bundling stop_summoning_prompt() together with it into a
single player_summon_check() function that can be used succinctly.

It also adds these checks to a number of evokers - sack of spiders, horn of
Geryon, box of beasts, and phantom mirror. In the process, it also fixes
some oddities with a couple of these. If the sack or box failed to create
a monster (but *not* the horn!) failed to produce a monster, it would not
take any time nor use up the evoker charge. But this was a real attempt,
so if the player had any chance at all of making broodmothers, they could
infinitely use the sack over deep water and fail for 'free' until rolling
high enough to get one.

Now, alas, an evoker failing to do anything still uses up the charge to
prevent these sort of shenanigans (but the game should be much better at
preventing the player from doing so in the vast majority of cases).

A few god abilities also get summon checks - Infernal Servant for space
and Infernal Legion / Blood for Blood for damaging auras.

Finally, fix Sphere of Battle not appearing while the player is maintaining
a damaging aura like Polar Vortex (which couldn't even hurt it!) while also
no longer spamming mysterious messages every turn that it wasn't able to
summon one due to a lack of space.