Don't let {!a} inscriptions stop spellclaws and similar (CarefulOdds)
The intent of the code was that various 'automatic' attacks like
spellclaws, spellmotor, and oni potion cleave, would never make attacks
that could harm allies or incur penance. Rather than prompting the player,
they simply wouldn't occur.

Which I think is still fine for unrand weapons with splash damage and the
like (where constant prompts could be intrusive for minor gain), but it
interacts oddly with the {!a} inscription, which now makes *any* attack
prompt (and thus prevents these automatic attacks from ocurring.)

One possible issue is that this inscription is used both for "My attack is
low-value, so don't waste actions making it" in which case there's no
reason anyone would want to skip *free* attacks. The other is often "This
weapon has *Rage on it and I don't want to berserk."

But inscriptions probably shouldn't have a 'gameplay' effect (where the
player can turn a mechanic on or off, but only by changing the inscription
on their weapon). So instead, we simply ignore that inscription now, which
is likely far more intuitive to everyone instead. (I know *I* didn't
expect this behaviour, until it was pointed out.)