Tweak some unidentified randart descriptors
Specifically, "crude" makes a randart sound like it's of lower quality,
and is one of the most obviously clashing with the given fanciful randart
tiles. It's been replaced in its multiple uses with a variety of different
fancier synonyms for "glowing", to match "glowing" being a common
non-artefact un-identified equipment descriptor.