Make Rending Blade much more reliable (and deal more damage)
While situations where rending blade was unable to find an attack path
when one clearly existed were a minority, they were still much too common,
to the point of making the spell feel fiddly and inconsistent. This commit
rewrites most of its pathing behavior entirely, in the hope of reducing
those confusing/frustrating scenarios to almost zero.

It uses a very heavy-handed approach to do so, iterating all possible paths
from its present position (and, if necessary, from all adjacent positions).
I'm not altogether happy with how heavyweight this is, but it's the best
way to ensure that it almost always does a thing if the player could see a
way for it to do a thing.

The blade can no longer 'store up' charge between turns to unleash in a
burst. This was intended to compensate for the 'unreliability' of the old
model, so that if it failed one turn, it might strike twice the next turn,
but was fairly unintuitive to players. Now it will either strike
immediately or not at all.

Finally, the blade leashes to the player like a phalanx beetle instead of
drifting around at a distance, semi-randomly. (The latter sometimes helped
it find paths in the old model, but is no longer helpful.)

In addition, the base damage of the blade has been raised, and the bonus
power per MP also raised.

I'm not altogether happy with this state, and feel the spell could possibly
use a deeper redesign, but there's no time to devise, implement, and test
one so close to feature freeze. And this is definitely a more usable state
for the spell than what currently exists.