Remove some worse-than-useless wizmode code for placing player ghosts
Playing a player ghost in wizmode would prompt the user to specify a
species and job for that ghost, but it had *already decided* by this point
in time what ghost it was placing and would just replace its listed species
and job with whatever you said - ie: make the ghost lie to you.

And then, on top of that, it always set their XL to 7 and HP to 20! (And
clobbered their name, too.)

So just remove all this code entirely. Now placing a player ghost will
simply place a random available ghost from the current floor. (It might be
nice to be able to specify what sort of ghost to pull up from the ones that
actually exist, but doing literally nothing is better than doing whatever
this was doing....)