Buff the Faerie dragon scales
Faerie dragon scales being a mildly-better acid dragon scales randart in
Depths is not very much of an incentive nor a distinguishing nature to it
outside of the 5% chances for slightly unique properties like Clarity,
+MP, or +HP. Its location is simply astonishingly late for characters to
change their body armour versus having enchanted other body armour options
up to that point. While power spiral concerns are always a bit of a
conscientious issue, the last adjustment to this was in b633e9c nearly
nine years ago, and we can afford another tweak to it after that timespan.

To make the fae stand out more compared to being a slightly modified
randart, it now always comes with a Hexes enhancer (suiting the
Enchantress it's found on, and not easily available with current enhancer
options), the aforementioned properties have been pushed to 10% chances
(nudged up to provide MP or HP in the range randart jewellery currently
gets for MP since 8d0d2ce), a new 10% chance for an invokable Invisibility
property, and a new enchantment range of +2 to +8 instead of +0 to +8.