Tweak the layout of a Lair ending
When I reworked due_jungle_book to give it the current undead theme, an
obsidian statue was added to guard a special necromancy-themed loot
item. And area in an adjacent chamber was made transparent to try to
reveal the statue before it would be reach via autoexplore. This is
important to avoid an autoexplore trap into a mesmerizing monster that
few players will be able to resist at this depth. However the revealed
area both wasn't perfect and was visually awkward. This commit removes
the transparent walls from the adjacent chamber and tweaks the forested
area near the statue so that the statue is always revealed regardless of
movement approach in a way where it can't affect the player.

Also the player now must actually step into a square where they can be
affected by the statue to reveal the loot item. Most of the loot in the
other chambers of this vault is revealed from relative safety, so it's
nice to have some variation by requiring a potentially dangerous action
in order to see the most thematic loot item.