Add Bane of Mortality
This bane gives a chance on any turn where the player is seriously injured,
but no non-trivial monsters are in sight, to summon hostile reapers to
claim their soul. (With a long cooldown whenever it happens.)

These are always summoned several tiles away from the player (to give them
time to react) and at low XL are also perma-slow (to allow the player to
flee). They always spawn with Warding (to keep 'kite a permaslow reaper for
100 years to hurl stones at it' from ever being a productive strategy.)

This is a bit of a reference to old Necro miscasts/death curses, but also:
the image of having to flee from an implacable but slow reaper when one is
vulnerable is very good flavor, in my opinion.

To prevent the player from having to do manual cleanup on reapers they've
parked or ran away from, clearing the bane itself immediately removes all
reapers created by its effects.