Re-add quasits with a new coat of paint
Now called drudes (a Germanic nightmare spirit) to trim off the DnD-isms,
these are once again fast minor demons that inflict strategic damage with
their melee - but now in the form of Doom.

They are part of the Summon Minor Demon list, but now also spawn in D
(since they can actually do something), taking a little weight from white
imps and shadow imps.

The numbers on their AF_DOOM are aimed with the expectation that a single
drude encounter is unlikely to result in a bane, but consecutive ones (if
handled inadequately) could - as well as leaving players more exposed to
the threat of another monster in the following commit.

Credit to regret-index for the edit to the quasit tile.