New status effect: Diminished Spells
This status reduces player spellpower by 50% (after stepdown and capping
power, so that the screen will always display half the power you ordinarily
had, instead of 'some opaquely variable amount less than normal'). On
monsters, it reduces their spell HD by 7 instead (to be more symmetric with
Empowered Spells raising it by 5).

Slated to be used by future monsters as a 'weak, but for spells' sort of
effect (as another avenue for interfering with spellcasting that is less
binary than silence or MP drain).