Make Enfeeble apply Diminished Spells to monsters instead of antimagic
The way Enfeeble applied antimagic to monsters was somewhat unintuitive and
poorly-conveyed. Antimagic is a status that cares heavily about being
repeatedly stacked, with its chance of blocking a spell being dependant on
how much duration remains. But Enfeeble, unlike every other player source
of antimagic, DOESN'T stack - each cast will roll the effect duration and
increase it only if the remaining duration is less. This means that
repeatedly casting the spell will, on average, do very little to impair a
monster's spellcasting any more than the first cast did. This isn't really
communicated by the game at all (and directly affects how the spell should
actually be used).

While an argument could be made that antimagic's internal mechanics could
also be changed or communicated better, applying Diminished Spells instead
has a more transparent effect that naturally does not stack (and mirrors
the Weak the spell already applies). And, importantly, is something that
can be done to the player without being direct MP drain, for a pending
monster version of the spell.