Remove internal zap power caps
These have served essentially no purpose for some time, as far as I can
tell, aside from being the source of occasional bugs. Monsters always
ignored them, and *spell* power caps supercede them for player usages
(although, if the zap for a spell had a lower cap than the spell, it would
effectively lower the spell power cap as well, which has resulted in bugs
in past).

It looks like the original reason for separating the two was to allow
spells where, at some point in their power curve, their range could be
increased *without* also increasing their power. No such spell has existed
for a very long time, and I cannot imagine we would choose to create
another, so it seems best to remove this.

(Though as always, with a refactor of this nature, it's possible this is
some localized effect I have overlooked, although I've tried to be
thorough.)