Make neqoxecs apply temp mutations instead of full ones (hellmonk)
While fear and player dissatisfaction over getting malmutated by neqoxecs
may be outsized compared to the real threat those mutations present, it
demonstrably has an effect on how players even *discuss* Extended (and,
for instance, the percieved importance of death form in Pan). Pan is due
for significant changes this version that may recontextualize a lot of
that anyway, but their use in 3-rune games is arguably awkward: a rare
result from a handful of summoners that is very unlikely to do anything to
you, but with that thin sliver of a chance of feeling very unfun to some
people.

I don't think there's a terrible lot of value in preserving the tiny chance
to get, say, berserkitis in Orc:2 (and I hope to do a series of things in
0.34 to make counterplay against bad mutations feel more agentive and
'push your luck' in general), so let's try making them inflict temporary
mutations instead.

Instead of Malmutate, they now cast Warp Body: a line-of-fire requiring
spell that inflicts a temp mutation and a small amount of damage (1d7).
But if the player already has 3 different tempmuts, they become Vitrified
instead of getting a 4th mutation. This keeps the mutation list from
getting too large to internalize, while also creating a more immediate
threat to endanger the player (especially if multiple neqoxecs are
cooperating together in Pan). Let's see how this feels!

-----------------

Warp Body spell icon by regret-index, combining ontoclasm's Ice Form tile
and NormalPerson7's Malmutate tile.