Expand the definition of 'trivial' monster threat level slightly
Even a simple dungeon yak yielded too much XP to ever be considered
trivial, even by xl 27. This is obviously nonsense (and now there is a
mechanical consequence to this nonsense), so let's expand the category
slightly.

(I think threat tiers require a much more detailed review/rework, which I
hope to do later in this version, but such a thing is out of scope for the
moment.)