Replace Horrible Things chance of tempmut with Doom
Instead of a chance to get a temp Will- mutation, now casting the spell
gives a small amount of doom instead (with a 1 in 4 chance of a higher
amount, along with the message).

The amount is fairly arbitrary and not intended to significantly nerf the
spell (but it's a bit hard to predict how bad it will be, given that
killing things with XXX will directly low the doom it gave the player). So
we'll have to see.