Revamp Trove fees using some new options
Troves are a theoretically exciting thing, and asking the player to
sacrifice something in exchange for a random reward can be interesting
gameplay, but many of the trove fees are troubled in practice.

Vital consumables are often too valuable for players to consider it
equitable to sacrifice them for an unknown reward, but more tricky is the
fact that the quantities asked for will in some cases exceed *how many of
those items generate for the player in the entire game*. Asking the player
to increase their challenge is one thing, but giving them tasks they cannot
complete (and might barely be worth it, even if they could) is
unsatisfying. And on the other end, options like 'item in a shop' are often
just "Gold (which will eventually be free), but complicated."

Instead, let's trim down the current options to focus on ones that are more
interesting, and give significant weight over to a new press-your-luck
option in the form of Banes - something the player can *always* pay, but
which still represents an immediate new threat to them.

-50% of Troves now ask the player to accept a pre-announced bane in order
 to enter. (The details of the bane are viewable when the portal is
 examined and I've excluded some of the more situational ones from the list
 of possibilities.)
-12.5% are piety-sacrifice troves (now using Ostracism instead, so that
 they function with more gods than before).
-~6.2% inflict heavy Draining on the player.
-~17% are split between Abyss/Slimy Rune and Horn of Geryone
-~4% require 2 acquirement scrolls
-~4% ask for an enchanted demon weapon (though any type will now suffice)
-~6% ask for enchanted dragon armour (with deeper types being rarer)

These numbers are subject to change in future (to be honest, I'm not sure
if the enchanted dragon armour is terrible interesting to keep, among the
other possibilities).

A subsequent commit will heavily revamp the *contents* of Troves as well.