Mildly tweak the monster version of Ozocubu's Refrigeration
It now no longer does friendly-fire, as this would make the Orb of Winter
version in particular hard for monsters to use (monster OTR has had this
same property for a very long time).

Slightly tweak power scaling so that it is marginally weaker at very high
HD (Fannar is largely unaffected).

Finally, remove a weirdness to its tracer that would sometimes prevent
monsters from ever casting the spell if the player's AC was high enough.
This did require 52 AC to shut off Fannar's Refrigeration, so it probably
didn't happen very often, but beam tracers in general ignore the player's
AC and it's still possible to do *some* damage against a high AC monster,
so this behavior seemed weird to me.