Fix krakens not flopping around on land (#4630)
A kraken can move into temporary deep water such as that created by the
summon forest spell and then be left on land when the water reverts
back. This should result in the kraken flopping around on land taking
damage. However, the check for this was checking if the monster's
habitat was HT_WATER and krakens are HT_DEEP_WATER now.