New good mutation: Lucky
This mutation gives a small chance (currently 2/4%) of any new floor item
the player encounters being upgraded to a randart of the same type - as
one of the most transparent and tractable implementations of a recurringly
suggested 'improve item luck' effect. It even announces the effect so that
players can realize it helped them!

I'm not sure exactly how generous this is in practice, since it excludes
monster items and shops (which are a large chunk of items one cares about),
but numbers are easily adjustable. And this feels like exactly the sort of
thing that could get some people more excited to drink !mutation for
reasons other than fixing problems.