Fix resizing the console while in a menu on windows (#4506)
Responding to console resize events in menus was disabled in 9746042 as
the way we were handling them lead to generating another resize event
resulting in an infinite loop. However, this results in the menu
becoming fairly unreadable until you exit it and go back in. With a
slight adjustment we can use the normal resize handling code in menus as
well which doesn't have this problem.