Nudge up devourer's traps
In line with 70c6c67 trying to make rarer and less consistent sources of
corrosion be more effective so they're more noticeable when they do
happen, and with the introduction of a corrosion effect doing more than 4
at a time, I'm nudging up devourer's traps to match the Corrosive Bolt
changes: 75% chance of -6 AC and slaying, as opposed to ~66% for -4.
They're far less directly effective (or anecdotally noticed) than the
other traps of their quartet, but are also inherently themed towards a
very late placement; as such, they can take the buff.